backend/app/api/auth.py
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from sqlalchemy.ext.asyncio import AsyncSession
from datetime import datetime, timedelta
from typing import Optional
import os
from jose import JWTError, jwt
from passlib.context import CryptContext

from ..database.session import get_db
from ..models.user import User

router = APIRouter()

# Security setup
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="api/auth/login")

# JWT configuration
SECRET_KEY = os.getenv("JWT_SECRET", "your-secret-key-for-development")
ALGORITHM = os.getenv("JWT_ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("JWT_ACCESS_TOKEN_EXPIRE_MINUTES", 30))

@router.post("/register")
async def register(
    username: str,
    email: str,
    password: str,
    full_name: Optional[str] = None,
    db: AsyncSession = Depends(get_db)
):
    """Register a new user"""
    # Placeholder for user registration logic
    return {"message": "User registration endpoint"}

@router.post("/login")
async def login(
    form_data: OAuth2PasswordRequestForm = Depends(),
    db: AsyncSession = Depends(get_db)
):
    """Login and receive JWT token"""
    # Placeholder for user authentication logic
    return {"message": "User login endpoint"}

@router.post("/refresh")
async def refresh_token(
    token: str = Depends(oauth2_scheme),
    db: AsyncSession = Depends(get_db)
):
    """Refresh JWT token"""
    # Placeholder for token refresh logic
    return {"message": "Token refresh endpoint"}

@router.post("/logout")
async def logout(
    token: str = Depends(oauth2_scheme)
):
    """Logout (invalidate token)"""
    # Placeholder for logout logic
    return {"message": "Logout endpoint"}

@router.get("/me")
async def get_current_user(
    token: str = Depends(oauth2_scheme),
    db: AsyncSession = Depends(get_db)
):
    """Get current user information"""
    # Placeholder for getting current user logic
    return {"message": "Get current user endpoint"}