backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
sqlalchemy>=2.0.20
aiosqlite>=0.19.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.1
google-generativeai>=0.3.0
TTS>=0.14.0
torch>=2.0.0
cloudinary>=1.33.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4